+++ packages/expo-router/src/link/Link.tsx
@@ -89,6 +89,7 @@ export function Redirect({ href, relativeToDirectory, withAnchor }: RedirectProp
 }
 
 /**
+<<<<<<< HEAD
  * Component that renders a link using [`href`](#href) to another route.
  * By default, it accepts children and wraps them in a `<Text>` component.
  *